requirements.txt
gradio
openai
python-dotenv
autogen
pyautogen
markdown
rich 